Fix memory leak, add set method
Added a destructor that frees the memory held by the m_vector field once the Vector object is out of scope. Also added a method that sets the vector object to point to a heap-allocated array of a given length. Will likely add smart pointers in the future, because I still feel like there could be a memory leak somewhere.